Add a timeout of 5 seconds to our Soup.SessionAsync object.
If retrieving the data is taking more than 5 seconds, probably it's
not comming in this iteration, so get a timeout and try again 3
seconds later (if there is a global connection available).

I've been (rarely) getting "busy" icons in the extension after waking
up from sleep-to-ram, even with the new querying of the NetworkManager
connection status. I believe this is happening if I suspend the
machine *exactly* when the Soup.SessionAsync object request the data
to Yahoo!, and so when it wakes up, the connection is not there. The
default timeout is huge, so I hope that by making it tiny, we don't
see this problem anymore.

Also, fixed the indentation of the NetworkManagerInterface.